[Issue-4412] refactor: Update UI logic for address suggestion
