+++ packages/extension-koni-ui/src/components/Field/AddressInputNew.tsx
@@ -15,7 +15,7 @@ import useGetChainInfo from '@subwallet/extension-koni-ui/hooks/screen/common/us
 import { cancelSubscription, saveRecentAccount, subscribeAccountsInputAddress } from '@subwallet/extension-koni-ui/messaging';
 import { RootState } from '@subwallet/extension-koni-ui/stores';
 import { ScannerResult, ThemeProps } from '@subwallet/extension-koni-ui/types';
-import { toShort } from '@subwallet/extension-koni-ui/utils';
+import { sortFuncAnalyzeAddress, toShort } from '@subwallet/extension-koni-ui/utils';
 import { isAddress } from '@subwallet/keyring';
 import { AutoComplete, Button, Icon, Input, ModalContext, Switch, SwQrScanner } from '@subwallet/react-ui';
 import CN from 'classnames';
@@ -199,10 +199,10 @@ function Component (props: Props, ref: ForwardedRef<AddressInputRef>): React.Rea
     }
 
     const result: AutoCompleteGroupItem[] = [];
-    const walletItems: AutoCompleteItem[] = [];
-    const contactItems: AutoCompleteItem[] = [];
-    const domainItems: AutoCompleteItem[] = [];
-    const recentItems: AutoCompleteItem[] = [];
+    const walletItems: AnalyzeAddress[] = [];
+    const contactItems: AnalyzeAddress[] = [];
+    const domainItems: AnalyzeAddress[] = [];
+    const recentItems: AnalyzeAddress[] = [];
 
     const genAutoCompleteItem = (responseOption: AnalyzeAddress): AutoCompleteItem => {
       return {
@@ -210,7 +210,7 @@ function Component (props: Props, ref: ForwardedRef<AddressInputRef>): React.Rea
         label: (
           <AddressSelectorItem
             address={responseOption.formatedAddress}
-            avatarValue={responseOption.proxyId}
+            avatarValue={responseOption.proxyId || responseOption.address}
             name={responseOption.displayName}
           />
         ),
@@ -227,30 +227,34 @@ function Component (props: Props, ref: ForwardedRef<AddressInputRef>): React.Rea
 
     responseOptions.forEach((ro) => {
       if (ro.analyzedGroup === AnalyzedGroup.WALLET) {
-        walletItems.push(genAutoCompleteItem(ro));
+        walletItems.push(ro);
       } else if (ro.analyzedGroup === AnalyzedGroup.CONTACT) {
-        contactItems.push(genAutoCompleteItem(ro));
+        contactItems.push(ro);
       } else if (ro.analyzedGroup === AnalyzedGroup.DOMAIN) {
-        domainItems.push(genAutoCompleteItem(ro));
+        domainItems.push(ro);
       } else if (ro.analyzedGroup === AnalyzedGroup.RECENT) {
-        recentItems.push(genAutoCompleteItem(ro));
+        recentItems.push(ro);
       }
     });
 
     if (walletItems.length) {
-      result.push(genAutoCompleteGroupItem(t('My wallet'), walletItems));
+      walletItems.sort(sortFuncAnalyzeAddress);
+      result.push(genAutoCompleteGroupItem(t('My wallet'), walletItems.map((i) => genAutoCompleteItem(i))));
     }
 
     if (contactItems.length) {
-      result.push(genAutoCompleteGroupItem(t('My contact'), contactItems));
+      contactItems.sort(sortFuncAnalyzeAddress);
+      result.push(genAutoCompleteGroupItem(t('My contact'), contactItems.map((i) => genAutoCompleteItem(i))));
     }
 
     if (domainItems.length) {
-      result.push(genAutoCompleteGroupItem(t('Domain name'), domainItems));
+      domainItems.sort(sortFuncAnalyzeAddress);
+      result.push(genAutoCompleteGroupItem(t('Domain name'), domainItems.map((i) => genAutoCompleteItem(i))));
     }
 
     if (recentItems.length) {
-      result.push(genAutoCompleteGroupItem(t('Recent'), recentItems));
+      recentItems.sort(sortFuncAnalyzeAddress);
+      result.push(genAutoCompleteGroupItem(t('Recent'), recentItems.map((i) => genAutoCompleteItem(i))));
     }
 
     return result;
